+++ CHANGELOG.md
@@ -10,6 +10,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 
 - We added a machine translated Tagalog locale (#1089).
 - We added additional keywords to all emojis to improve searching (#1090), but also translated all of those keywords, so that works in all supported locales (#639).
+- We added much more robust menu suggestions for locales and locale parts (#1099).
 
 ### Changed